+++ e2e/model_test.go
@@ -24,9 +24,6 @@ func TestRegisterModel(t *testing.T) {
 	appClient.RegisterComputePlan(client.DefaultComputePlanOptions())
 	appClient.RegisterTasks(client.DefaultTrainTaskOptions())
 
-	plan := appClient.GetComputePlan("cp")
-	require.EqualValues(t, 1, plan.TaskCount)
-
 	appClient.StartTask(client.DefaultTrainTaskRef)
 	registeredModel := appClient.RegisterModel(client.DefaultModelOptions())