+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Build images with go 1.19
 - Add a `Transient` field to the task inputs
 
+### Removed
+
+- Task counts by status from ComputePlan responses
+
 ## [0.21.0] - 2022-08-01
 
 ### Added